refactor(estimates): extract useEstimateAI sub-hook (Phase 1b)
All AI/scope state and handlers extracted from useEstimateForm into
useEstimateAI. Main hook reduced from 895 to ~530 lines; AI logic
encapsulated with callback pattern for cross-hook state updates.
